Editing & Organising of Code & Edited SecurityConfig
- Changed Entity Namings and moved entity files out of folders
- Edited Security Config files to permit all apis and swagger
