+++ README.md
@@ -1,20 +1,16 @@
 # Authenticket
 - SMU CS203 Collaborative Software Development (AY23/24)
 
-##Backend
+## Backend
 
-###To run backend in IntelliJ
-Firstly, reload pom.xml file to reimport dependencies
-
-Reimporting dependencies:
+### Reimporting dependencies
 	1. Open your project in IntelliJ IDEA.
 	2. Locate the `pom.xml` file in the root directory of your project.
 	3. Right-click on the `pom.xml` file.
 	4. From the context menu, hover over the option "Maven"
 	5. Click on "Reload Object"
 
-2. If everything works, you should be able to click on run button on top right corner to 
+### Take note: you will need to reload pom.xml file every time you make changes to your pom.xml to ensure that your project reflects the changes.
 
-###Take note: you will need to reload pom.xml file every time you make changes to your pom.xml to ensure that your project reflects the changes.
-
-###Swagger API URL: http://localhost:8080/swagger-ui/index.html#/
+## API Documention
+### Swagger API URL: <http://localhost:8080/swagger-ui/index.html#/>
