--- backend/src/main/java/com/authenticket/authenticket/controller/UserController.java
@@ -1,18 +1,0 @@
-package com.authenticket.authenticket.controller;
-
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@CrossOrigin
-@RequestMapping("/user")
-
-public class UserController {
-    @GetMapping("/test")
-    public String test() {
-        return "test successful";
-    }
-
-}
