+++ backend/src/main/java/com/authenticket/authenticket/controller/user/UserController.java
@@ -13,33 +13,9 @@ import java.util.Optional;
     @RequestMapping(path = "/registration")
 
     public class UserController {
-
-//        @Autowired
-//        private PublisherRepository publisherRepository;
-
-//        @PostMapping("/create")
-//        public Publisher create(@RequestBody Publisher publisher) {
-//            return publisherRepository.save(publisher);
-//        }
-//
         @GetMapping("/test")
         public String test() {
           return "test successful";
         }
-//
-//        @GetMapping("/{id}")
-//        public Optional<Publisher> findOneById(@PathVariable String id) {
-//            return publisherRepository.findById(id);
-//        }
-//
-//        @PutMapping("/update")
-//        public Publisher update(@RequestBody Publisher publisher) {
-//            return publisherRepository.save(publisher);
-//        }
-//
-//        @DeleteMapping("/delete/{id}")
-//        public void deleteById(@PathVariable String id) {
-//            publisherRepository.deleteById(id);
-//        }
 
     }
