+++ backend/src/main/java/com/authenticket/authenticket/model/EventOrganiser.java
@@ -1,4 +1,4 @@
 package com.authenticket.authenticket.model;
 
-public class EventOrganiserModel {
+public class EventOrganiser {
 }
