+++ backend/src/main/java/com/authenticket/authenticket/model/TicketCategory.java
@@ -1,4 +1,4 @@
 package com.authenticket.authenticket.model;
 
-public class TicketCategoryModel {
+public class TicketCategory {
 }
