--- backend/src/main/java/com/authenticket/authenticket/model/UserModel.java
@@ -1,4 +1,0 @@
-package com.authenticket.authenticket.model;
-
-public class UserModel {
-}
