+++ backend/src/main/java/com/authenticket/authenticket/service/AdminService.java
@@ -1,0 +1,4 @@
+package com.authenticket.authenticket.service;
+
+public class AdminService {
+}
