+++ backend/src/main/java/com/authenticket/authenticket/service/TicketCategoryService.java
@@ -1,0 +1,4 @@
+package com.authenticket.authenticket.service;
+
+public class TicketCategoryService {
+}
