+++ backend/src/main/java/com/authenticket/authenticket/service/TicketService.java
@@ -1,0 +1,4 @@
+package com.authenticket.authenticket.service;
+
+public class TicketService {
+}
