+++ backend/src/main/java/com/authenticket/authenticket/service/UserService.java
@@ -1,0 +1,5 @@
+package com.authenticket.authenticket.service;
+
+public class UserService {
+
+}
