--- backend/src/main/java/com/authenticket/authenticket/service/impl/EventOrganiserService.java
@@ -1,4 +1,0 @@
-package com.authenticket.authenticket.service.impl;
-
-public class EventOrganiserService {
-}
