--- backend/src/main/java/com/authenticket/authenticket/service/impl/TicketService.java
@@ -1,4 +1,0 @@
-package com.authenticket.authenticket.service.impl;
-
-public class TicketService {
-}
