--- backend/src/main/java/com/authenticket/authenticket/service/impl/UserService.java
@@ -1,5 +1,0 @@
-package com.authenticket.authenticket.service.impl;
-
-public class UserService {
-
-}
